pyproject.toml
[project]
name = "ecsr-cot"
version = "0.1.0"
description = "Error-Correcting State Recurrence CoT experiments"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.1.0",
  "transformers>=4.38.0",
  "datasets>=2.16.0",
  "hydra-core>=1.3.2",
  "omegaconf>=2.3.0",
  "wandb>=0.16.0",
  "optuna>=3.5.0",
  "numpy>=1.23.0",
  "pandas>=1.5.0",
  "matplotlib>=3.7.0",
  "seaborn>=0.13.0",
  "scipy>=1.10.0",
  "accelerate>=0.27.0",
  "sentencepiece>=0.1.99"
]
